// apps/web/src/generated/api/models/userProfileResponseDto.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * ThreadNest API
 * OpenAPI spec version: 1.0.0
 */

export interface UserProfileResponseDto {
  /** User ID */
  userId: string
  /** Unique username */
  username: string
  /**
     * Display name
     * @nullable
     */
  displayName: string | null
  /**
     * Bio
     * @nullable
     */
  bio: string | null
  /**
     * Avatar URL
     * @nullable
     */
  avatarUrl: string | null
  /** When the profile was created */
  createdAt: string
}